Atualizar o ex2.c

+++ ex2.c
@@ -5,7 +5,7 @@ Enzo Vega Grozdas - AED2 - 2AN - Ex 2
 *******************************************************************************/
 #include <stdio.h>
 
-int divisao(int n1, int n2)
+double divisao(int n1, int n2)
 {
     return (n1 / n2);
 }
